pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "batch_allocations"
version = "1.0.0"
description = "This system follows along the examples in Architecture Patterns with Python"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ricardo Zambrano", email = "rzambrano@gmail.com"}
    ]

dependencies = [
    "sqlalchemy",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]

dev = [
    "pytest",
    "pytest-cov",
    "tox",
    "black",
    "ruff",
    "mypy",
]

web = [
    "flask>=3.0",
    "requests>=2.31",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
# include = ["batch_allocations*"]
# namespaces = false

[tool.setuptools.package-data]
"batch_allocations" = ["py.typed"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = [
    "--cov", 
    "--strict-markers", # makes pytest fail if you use an undefined marker
    "-v",               # add verbose output
    # "--setup-show",     # add fixture setup/teardown info
    ]
testpaths = [
    "test"
]
xfail_strict = true # treats tests marked as "expected to fail" (xfail) as a failure if they actually pass

# markers = [] # To register markers

[tool.coverage.run]
branch = true
source = ["src"]  

[tool.coverage.report]
show_missing = true
skip_covered = true

[tool.mypy]
python_version = "3.11"  
warn_unused_configs = true
show_error_context = true
pretty = true
namespace_packages = true
check_untyped_defs = true
explicit_package_bases = true  
mypy_path = "src" 
exclude = [
    '\.venv',
    '\.tox',
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "N",   # pep8-naming
]
ignore = []